requirements.txt
langchain
langchain-openai
langchain-community
langchain-chroma
chromadb
sentence-transformers
rank-bm25
pypdf
ragas
datasets
openai
streamlit
python-dotenv
pandas
